+++ sw.js
@@ -1,4 +1,4 @@
-var CACHE_VERSION = 135;
+var CACHE_VERSION = 136;
 var CACHE_NAME = 'nyhome-v' + CACHE_VERSION;
 
 var APP_SHELL = [